Refactor CubeDemo: Modularize game structure and enhance documentation
- Introduced CubeDemoGame class to encapsulate game logic and lifecycle management.
- Created CubeDemoScene for scene setup, including cube creation and camera configuration.
- Added FreeCamera class to manage camera movement and rotation.
- Implemented CubeDemoSystems for system registration and game logic execution.
- Enhanced documentation with detailed comments for clarity on class responsibilities and functionality.
- Updated CMakeLists.txt to include new source files for better organization.
