+++ engine/include/app/Game.h
@@ -2,6 +2,12 @@
 
 #include <app/GameContexts.h>
 
+//=============================================================================
+// Game
+//
+// Base interface for application-specific configuration and engine callbacks.
+// Games override lifecycle hooks to register systems and react to platform events.
+//=============================================================================
 class Game
 {
 public:
