+++ engine/include/app/GameContexts.h
@@ -14,23 +14,47 @@
 class Engine;
 class EngineSchedule;
 
+//=============================================================================
+// GameConfigureContext
+//
+// Provides mutable engine configuration before the Engine is constructed.
+// Used by a Game to choose startup settings and service options.
+//=============================================================================
 struct GameConfigureContext
 {
     EngineConfig& Config;
 };
 
+//=============================================================================
+// GameStartupContext
+//
+// Provides engine and configuration access during game startup.
+// Used after engine initialization and before normal frame processing begins.
+//=============================================================================
 struct GameStartupContext
 {
     Engine& EngineInstance;
     EngineConfig& Config;
 };
 
+//=============================================================================
+// GameShutdownContext
+//
+// Provides engine and configuration access during game shutdown.
+// Used for game-owned cleanup while engine services are still reachable.
+//=============================================================================
 struct GameShutdownContext
 {
     Engine& EngineInstance;
     EngineConfig& Config;
 };
 
+//=============================================================================
+// SystemRegisterContext
+//
+// Provides access to the engine schedule during system registration.
+// Used by a Game to add systems and declare their execution dependencies.
+//=============================================================================
 struct SystemRegisterContext
 {
     Engine& EngineInstance;
@@ -38,6 +62,12 @@ struct SystemRegisterContext
     EngineSchedule& Schedule;
 };
 
+//=============================================================================
+// PlatformEventContext
+//
+// Wraps an SDL platform event with engine state and a handled flag.
+// Used by a Game to consume window, input, and platform messages.
+//=============================================================================
 struct PlatformEventContext
 {
     Engine& EngineInstance;
@@ -46,6 +76,12 @@ struct PlatformEventContext
     bool Handled = false;
 };
 
+//=============================================================================
+// FixedLogicContext
+//
+// Provides fixed-tick simulation state before physics runs.
+// Used for deterministic gameplay logic that advances on fixed time steps.
+//=============================================================================
 struct FixedLogicContext
 {
     Engine& EngineInstance;
@@ -57,6 +93,12 @@ struct FixedLogicContext
     std::span<Registry*> ActiveRegistries;
 };
 
+//=============================================================================
+// PhysicsContext
+//
+// Provides fixed-tick simulation state for physics systems.
+// Used for collision, dynamics, and other physics-authoritative updates.
+//=============================================================================
 struct PhysicsContext
 {
     Engine& EngineInstance;
@@ -68,6 +110,12 @@ struct PhysicsContext
     std::span<Registry*> ActiveRegistries;
 };
 
+//=============================================================================
+// PostFixedContext
+//
+// Provides fixed-tick simulation state after physics has completed.
+// Used for transform propagation and follow-up logic that depends on physics.
+//=============================================================================
 struct PostFixedContext
 {
     Engine& EngineInstance;
@@ -104,6 +152,12 @@ struct FrameUpdateContext
     std::span<Registry*> ActiveRegistries;
 };
 
+//=============================================================================
+// RenderExtractContext
+//
+// Provides presentation-time state and double-buffered render packets.
+// Used to extract renderable data from simulation state for the renderer.
+//=============================================================================
 struct RenderExtractContext
 {
     Engine& EngineInstance;
@@ -117,6 +171,12 @@ struct RenderExtractContext
     std::span<Registry*> ActiveRegistries;
 };
 
+//=============================================================================
+// AudioContext
+//
+// Provides presentation-time state for audio systems.
+// Used to update audio playback from current input and world registry views.
+//=============================================================================
 struct AudioContext
 {
     Engine& EngineInstance;
@@ -128,6 +188,12 @@ struct AudioContext
     std::span<Registry*> ActiveRegistries;
 };
 
+//=============================================================================
+// EndFrameContext
+//
+// Provides final per-frame state before the engine advances frame resources.
+// Used for cleanup, lifecycle-only work, and end-of-frame bookkeeping.
+//=============================================================================
 struct EndFrameContext
 {
     Engine& EngineInstance;
