+++ example/CubeDemo/CMakeLists.txt
@@ -1,2 +1,8 @@
-add_executable(CubeDemo main.cpp)
+add_executable(CubeDemo
+    main.cpp
+    CubeDemoGame.cpp
+    CubeDemoScene.cpp
+    CubeDemoSystems.cpp
+    FreeCamera.cpp
+)
 target_link_libraries(CubeDemo PRIVATE engine)
