+++ example/CubeDemo/CubeDemoSystems.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "CubeDemoScene.h"
+#include "FreeCamera.h"
+
+#include <app/EngineSchedule.h>
+#include <world/registry/Registry.h>
+
+void RegisterCubeDemoSystems(EngineSchedule& schedule,
+                             Registry*& registry,
+                             FreeCamera& freeCamera,
+                             DemoScene& scene);
