+++ example/CubeDemo/main.cpp
@@ -1,447 +1,7 @@
+#include "CubeDemoGame.h"
+
 #include <app/Application.h>
-#include <app/Engine.h>
-#include <app/Game.h>
 #include <core/config/EngineConfig.h>
-#include <core/logging/LoggingProvider.h>
-#include <debug/DebugLogSink.h>
-#include <debug/DebugService.h>
-#include <graphics/vulkan/Renderer.h>
-#include <graphics/vulkan/VulkanBufferService.h>
-#include <graphics/vulkan/VulkanFrameService.h>
-#include <graphics/vulkan/VulkanInstanceService.h>
-#include <graphics/vulkan/VulkanSwapchainService.h>
-#include <input/InputFrame.h>
-#include <math/Quat.h>
-#include <platform/SdlWindow.h>
-#include <platform/SdlWindowService.h>
-#include <render/Camera.h>
-#include <render/Material.h>
-#include <render/MeshRendererComponent.h>
-#include <render/MeshService.h>
-#include <render/RenderQueue.h>
-#include <world/entity/EntityRegistry.h>
-#include <world/registry/Registry.h>
-#include <world/transform/TransformStore.h>
-#include <zone/DefaultZoneBuilder.h>
-
-#ifdef SENCHA_ENABLE_DEBUG_UI
-#include <debug/ConsolePanel.h>
-#include <debug/IDebugPanel.h>
-#include <debug/ImGuiDebugOverlay.h>
-#include <debug/TimingPanel.h>
-#include <imgui.h>
-#endif
-
-#include <SDL3/SDL.h>
-
-#include <array>
-#include <cstdio>
-#include <memory>
-
-namespace
-{
-    struct DemoScene
-    {
-        EntityId Camera;
-        EntityId CenterCube;
-        MeshHandle CubeMesh;
-        MaterialHandle Red;
-        MaterialHandle Green;
-        MaterialHandle Blue;
-    };
-
-    struct FreeCamera
-    {
-        EntityId Entity;
-        float Yaw = 0.0f;
-        float Pitch = 0.0f;
-        float MoveSpeed = 4.0f;
-        float FastMultiplier = 4.0f;
-        float MouseSensitivity = 0.0025f;
-        bool LookHeld = false;
-
-        void UpdateLook(const InputFrame& input)
-        {
-            LookHeld = input.IsMouseButtonDown(SDL_BUTTON_RIGHT);
-            if (!LookHeld) return;
-            Yaw -= input.MouseDeltaX * MouseSensitivity;
-            Pitch -= input.MouseDeltaY * MouseSensitivity;
-            Pitch = SDL_clamp(Pitch, -1.5f, 1.5f);
-        }
-
-        void TickFixed(const InputFrame& input,
-                       TransformStore<Transform3f>& transforms,
-                       float fixedDt)
-        {
-            Transform3f* transform = transforms.TryGetLocalMutable(Entity);
-            if (transform == nullptr) return;
-
-            Vec3d move = Vec3d::Zero();
-            if (input.IsKeyDown(SDL_SCANCODE_W)) move += Vec3d::Forward();
-            if (input.IsKeyDown(SDL_SCANCODE_S)) move += Vec3d::Backward();
-            if (input.IsKeyDown(SDL_SCANCODE_D)) move += Vec3d::Right();
-            if (input.IsKeyDown(SDL_SCANCODE_A)) move += Vec3d::Left();
-            if (input.IsKeyDown(SDL_SCANCODE_E)) move += Vec3d::Up();
-            if (input.IsKeyDown(SDL_SCANCODE_Q)) move += Vec3d::Down();
-
-            ApplyRotation(transforms);
-
-            if (move.SqrMagnitude() > 0.0f)
-            {
-                move = move.Normalized();
-                const float speed = MoveSpeed
-                    * (input.IsKeyDown(SDL_SCANCODE_LSHIFT) ? FastMultiplier : 1.0f);
-                transform->Position += transform->Rotation.RotateVector(move) * (speed * fixedDt);
-            }
-        }
-
-        void ApplyRotation(TransformStore<Transform3f>& transforms) const
-        {
-            Transform3f* transform = transforms.TryGetLocalMutable(Entity);
-            if (transform == nullptr) return;
-
-            transform->Rotation =
-                Quatf::FromAxisAngle(Vec3d::Up(), Yaw)
-                * Quatf::FromAxisAngle(Vec3d::Right(), Pitch);
-        }
-    };
-
-    TransformStore<Transform3f>& Transforms(Registry& registry)
-    {
-        return registry.Components.Get<TransformStore<Transform3f>>();
-    }
-
-    void AddCube(Registry& registry,
-                 MeshHandle mesh,
-                 MaterialHandle material,
-                 const Vec3d& position,
-                 const Vec3d& scale)
-    {
-        EntityId entity = CreateDefaultEntity(registry, Transform3f(position, Quatf::Identity(), scale));
-        AddDefaultMeshRenderer(registry, entity, mesh, material);
-    }
-
-    DemoScene CreateScene(Registry& registry,
-                          MeshService& meshes,
-                          MaterialStore& materials,
-                          FreeCamera& freeCamera)
-    {
-        DemoScene scene;
-        scene.CubeMesh = meshes.Create(MeshPrimitives::BuildCube(1.0f));
-        scene.Red = materials.Create(Material{
-            .Pass = ShaderPassId::ForwardOpaque,
-            .BaseColor = Vec4(1.0f, 0.15f, 0.1f, 1.0f),
-        });
-        scene.Green = materials.Create(Material{
-            .Pass = ShaderPassId::ForwardOpaque,
-            .BaseColor = Vec4(0.1f, 0.85f, 0.45f, 1.0f),
-        });
-        scene.Blue = materials.Create(Material{
-            .Pass = ShaderPassId::ForwardOpaque,
-            .BaseColor = Vec4(0.2f, 0.45f, 1.0f, 1.0f),
-        });
-
-        scene.CenterCube = CreateDefaultEntity(registry, Transform3f(
-            Vec3d(0.0f, 0.0f, 0.0f), Quatf::Identity(), Vec3d::One()));
-        AddDefaultMeshRenderer(registry, scene.CenterCube, scene.CubeMesh, scene.Red);
-
-        AddCube(registry, scene.CubeMesh, scene.Green, Vec3d(-2.2f, 0.0f, -1.0f), Vec3d::One());
-        AddCube(registry, scene.CubeMesh, scene.Blue, Vec3d(2.2f, 0.0f, -1.0f), Vec3d(0.75f, 1.5f, 0.75f));
-        AddCube(registry, scene.CubeMesh, scene.Green, Vec3d(0.0f, -1.4f, -3.0f), Vec3d(8.0f, 0.15f, 8.0f));
-
-        scene.Camera = CreateDefaultEntity(registry, Transform3f(
-            Vec3d(0.0f, 1.0f, 5.0f), Quatf::Identity(), Vec3d::One()));
-        AddDefaultCamera(registry, scene.Camera, CameraComponent{
-            .Projection = ProjectionKind::Perspective,
-            .FovYRadians = 1.22173048f,
-            .NearPlane = 0.1f,
-            .FarPlane = 1000.0f,
-        });
-
-        freeCamera.Entity = scene.Camera;
-        return scene;
-    }
-
-#ifdef SENCHA_ENABLE_DEBUG_UI
-    class CubeDemoPanel : public IDebugPanel
-    {
-    public:
-        CubeDemoPanel(RenderQueue& queue,
-                      CameraRenderData& camera,
-                      FreeCamera& freeCamera,
-                      Registry& registry,
-                      DemoScene& scene)
-            : Queue(queue)
-            , Camera(camera)
-            , FreeCam(freeCamera)
-            , RegistryInstance(registry)
-            , Scene(scene)
-        {
-        }
-
-        void Draw() override
-        {
-            ImGui::Begin("Cube Demo");
-            ImGui::Text("Right mouse: look");
-            ImGui::Text("WASD: move, Q/E: down/up, Shift: fast");
-            ImGui::Separator();
-            ImGui::Text("Opaque queue: %zu", Queue.Opaque().size());
-            ImGui::Text("Camera entity: %u", Camera.Entity.Index);
-            ImGui::DragFloat("Move speed", &FreeCam.MoveSpeed, 0.1f, 0.1f, 50.0f);
-            ImGui::DragFloat("Mouse sensitivity", &FreeCam.MouseSensitivity, 0.0001f, 0.0001f, 0.02f);
-
-            if (Transform3f* cube = Transforms(RegistryInstance).TryGetLocalMutable(Scene.CenterCube))
-                ImGui::DragFloat3("Center cube position", &cube->Position.X, 0.05f);
-
-            ImGui::End();
-        }
-
-    private:
-        RenderQueue& Queue;
-        CameraRenderData& Camera;
-        FreeCamera& FreeCam;
-        Registry& RegistryInstance;
-        DemoScene& Scene;
-    };
-#endif
-
-    struct FreeCameraMovementSystem
-    {
-        FreeCameraMovementSystem(Registry*& registry, FreeCamera& freeCamera)
-            : RegistryInstance(registry)
-            , FreeCam(freeCamera)
-        {
-        }
-
-        void FixedLogic(FixedLogicContext& ctx)
-        {
-            if (RegistryInstance == nullptr) return;
-            FreeCam.TickFixed(
-                ctx.Input,
-                Transforms(*RegistryInstance),
-                static_cast<float>(ctx.Time.DeltaSeconds));
-        }
-
-        Registry*& RegistryInstance;
-        FreeCamera& FreeCam;
-    };
-
-    struct CubeSpinSystem
-    {
-        CubeSpinSystem(Registry*& registry, DemoScene& scene)
-            : RegistryInstance(registry)
-            , Scene(scene)
-        {
-        }
-
-        void FixedLogic(FixedLogicContext& ctx)
-        {
-            if (RegistryInstance == nullptr) return;
-            if (Transform3f* cube = Transforms(*RegistryInstance).TryGetLocalMutable(Scene.CenterCube))
-                cube->Rotation *= Quatf::FromAxisAngle(
-                    Vec3d::Up(),
-                    static_cast<float>(ctx.Time.DeltaSeconds));
-        }
-
-        Registry*& RegistryInstance;
-        DemoScene& Scene;
-    };
-
-    struct FreeCameraLookSystem
-    {
-        FreeCameraLookSystem(Registry*& registry, FreeCamera& freeCamera)
-            : RegistryInstance(registry)
-            , FreeCam(freeCamera)
-        {
-        }
-
-        void FrameUpdate(FrameUpdateContext& ctx)
-        {
-            if (RegistryInstance == nullptr) return;
-            FreeCam.UpdateLook(ctx.Input);
-            FreeCam.ApplyRotation(Transforms(*RegistryInstance));
-        }
-
-        Registry*& RegistryInstance;
-        FreeCamera& FreeCam;
-    };
-
-    struct MouseTraceSystem
-    {
-        explicit MouseTraceSystem(FreeCamera& freeCamera)
-            : FreeCam(freeCamera)
-        {
-        }
-
-        void FrameUpdate(FrameUpdateContext& ctx)
-        {
-            TraceHistory[TraceWrite] = TraceSample{
-                .Dt = ctx.WallDeltaSeconds,
-                .Mdx = ctx.Input.MouseDeltaX,
-                .Mdy = ctx.Input.MouseDeltaY,
-                .Yaw = FreeCam.Yaw,
-                .Pitch = FreeCam.Pitch,
-                .LookHeld = ctx.Input.IsMouseButtonDown(SDL_BUTTON_RIGHT),
-            };
-            TraceWrite = (TraceWrite + 1) % kTraceCapacity;
-            if (TraceCount < kTraceCapacity) ++TraceCount;
-
-            int numKeys = 0;
-            const bool* sdlKeys = SDL_GetKeyboardState(&numKeys);
-            const bool f2Down = sdlKeys != nullptr
-                && SDL_SCANCODE_F2 < numKeys
-                && sdlKeys[SDL_SCANCODE_F2];
-            if (f2Down && !F2WasDown)
-                DumpTrace();
-            F2WasDown = f2Down;
-        }
-
-        void DumpTrace()
-        {
-            std::fprintf(stderr, "---- mouse trace (last %zu frames) ----\n", TraceCount);
-            const size_t start = (TraceWrite + kTraceCapacity - TraceCount) % kTraceCapacity;
-            for (size_t i = 0; i < TraceCount; ++i)
-            {
-                const TraceSample& s = TraceHistory[(start + i) % kTraceCapacity];
-                std::fprintf(stderr, "[%02zu] dt=%.4f mdx=%+7.2f mdy=%+7.2f yaw=%+.4f pitch=%+.4f look=%d\n",
-                    i, s.Dt, s.Mdx, s.Mdy, s.Yaw, s.Pitch, s.LookHeld ? 1 : 0);
-            }
-            std::fflush(stderr);
-        }
-
-        struct TraceSample
-        {
-            double Dt = 0.0;
-            float Mdx = 0.0f;
-            float Mdy = 0.0f;
-            float Yaw = 0.0f;
-            float Pitch = 0.0f;
-            bool LookHeld = false;
-        };
-
-        static constexpr size_t kTraceCapacity = 120;
-        std::array<TraceSample, kTraceCapacity> TraceHistory{};
-        size_t TraceWrite = 0;
-        size_t TraceCount = 0;
-        bool F2WasDown = false;
-        FreeCamera& FreeCam;
-    };
-
-    class CubeDemoGame final : public Game
-    {
-    public:
-        void OnStart(GameStartupContext& ctx) override
-        {
-            Engine& engine = ctx.EngineInstance;
-            ServiceHost& services = engine.Services();
-            LoggingProvider& logging = services.GetLoggingProvider();
-            DebugService& debug = services.Get<DebugService>();
-            DebugLogSink& debugLog = debug.GetLogSink();
-            auto& buffers = services.Get<VulkanBufferService>();
-
-            Meshes = std::make_unique<MeshService>(logging, buffers);
-            DemoRegistry = &CreateDefault3DZone(
-                engine.Zones(), ZoneId{ 1 },
-                ZoneParticipation{ .Visible = true, .Logic = true });
-
-            Demo = CreateScene(*DemoRegistry, *Meshes, Materials, FreeCam);
-
-            engine.AddDefaultMeshRenderFeature(*Meshes, Materials);
-
-#ifdef SENCHA_ENABLE_DEBUG_UI
-            auto& windows = services.Get<SdlWindowService>();
-            SdlWindow* window = windows.GetPrimaryWindow();
-            auto& instance = services.Get<VulkanInstanceService>();
-            auto& frames = services.Get<VulkanFrameService>();
-
-            auto debugOverlay =
-                std::make_unique<ImGuiDebugOverlay>(debug, *window, instance, frames);
-            debugOverlay->AddPanel<ConsolePanel>(debugLog);
-            debugOverlay->AddPanel<TimingPanel>(engine.Timing());
-            debugOverlay->AddPanel<CubeDemoPanel>(
-                engine.GetRenderQueue(), engine.GetCameraData(), FreeCam, *DemoRegistry, Demo);
-            DebugOverlay = debugOverlay.get();
-            auto& renderer = services.Get<Renderer>();
-            renderer.AddFeature(std::move(debugOverlay));
-#else
-            (void)debugLog;
-            debug.Open();
-#endif
-
-            std::printf("Sencha Cube Demo\n");
-            std::printf("  Right mouse + move: look\n");
-            std::printf("  WASD: move, Q/E: down/up, Shift: fast\n");
-            std::printf("  F1: pause simulation (timescale 0)\n");
-            std::printf("  `: debugger when built with SENCHA_ENABLE_DEBUG_UI=ON\n");
-            std::printf("  Escape: quit\n");
-        }
-
-        void OnRegisterSystems(SystemRegisterContext& ctx) override
-        {
-            ctx.Schedule.Register<MouseTraceSystem>(FreeCam);
-            ctx.Schedule.Register<FreeCameraLookSystem>(DemoRegistry, FreeCam);
-            ctx.Schedule.Register<FreeCameraMovementSystem>(DemoRegistry, FreeCam);
-            ctx.Schedule.Register<CubeSpinSystem>(DemoRegistry, Demo);
-        }
-
-        void OnPlatformEvent(PlatformEventContext& ctx) override
-        {
-#ifdef SENCHA_ENABLE_DEBUG_UI
-            if (DebugOverlay != nullptr && DebugOverlay->ProcessSdlEvent(ctx.Event))
-                ctx.Handled = true;
-#else
-            (void)ctx;
-#endif
-            if (ctx.Handled)
-                return;
-
-            if (ctx.Event.type == SDL_EVENT_MOUSE_BUTTON_DOWN
-                && ctx.Event.button.button == SDL_BUTTON_RIGHT)
-            {
-                SetRelativeMouseMode(ctx.EngineInstance, true);
-            }
-            else if (ctx.Event.type == SDL_EVENT_MOUSE_BUTTON_UP
-                && ctx.Event.button.button == SDL_BUTTON_RIGHT)
-            {
-                SetRelativeMouseMode(ctx.EngineInstance, false);
-            }
-            else if (ctx.Event.type == SDL_EVENT_WINDOW_FOCUS_LOST)
-            {
-                SetRelativeMouseMode(ctx.EngineInstance, false);
-            }
-        }
-
-        void OnShutdown(GameShutdownContext& ctx) override
-        {
-#ifdef SENCHA_ENABLE_DEBUG_UI
-            DebugOverlay = nullptr;
-#endif
-            SetRelativeMouseMode(ctx.EngineInstance, false);
-        }
-
-    private:
-        void SetRelativeMouseMode(Engine& engine, bool enabled)
-        {
-            SdlWindow* window = engine.Services().Get<SdlWindowService>().GetPrimaryWindow();
-            if (window == nullptr || window->GetHandle() == nullptr)
-                return;
-
-            if (SDL_GetWindowRelativeMouseMode(window->GetHandle()) == enabled)
-                return;
-
-            SDL_SetWindowRelativeMouseMode(window->GetHandle(), enabled);
-        }
-
-        Registry* DemoRegistry = nullptr;
-        MaterialStore Materials;
-        std::unique_ptr<MeshService> Meshes;
-        FreeCamera FreeCam;
-        DemoScene Demo;
-#ifdef SENCHA_ENABLE_DEBUG_UI
-        ImGuiDebugOverlay* DebugOverlay = nullptr;
-#endif
-    };
-}
 
 int main(int argc, char** argv)
 {
